[CO-346] Move Mnemonic-stuff to BIP39.hs
Moves Mnemonic-phrase- and mneomnic-sentence- related things from
Dictionary.hs to BIP39.hs. This way, they can access the EntropyError we
soon will introduce.
