+++ src/Crypto/Encoding/BIP39.hs
@@ -62,6 +62,7 @@ import           Basement.String (String)
 import qualified Basement.String as String
 import           Basement.Nat
 import qualified Basement.Sized.List as ListN
+import           Basement.Sized.List (ListN)
 import           Basement.NormalForm
 import           Basement.Compat.Typeable
 import           Basement.Numerical.Number (IsIntegral(..))
@@ -72,7 +73,8 @@ import           Foundation.Check
 import           Control.Monad (replicateM)
 import           Data.Bits
 import           Data.Maybe (fromMaybe)
-import           Data.List (reverse)
+import           Data.List (reverse, intersperse)
+import           Data.Kind (Constraint)
 import           Data.ByteArray (ByteArrayAccess, ByteArray)
 import qualified Data.ByteArray as BA
 import           Data.ByteString (ByteString)
@@ -81,6 +83,7 @@ import qualified Data.ByteString as BS
 import           Data.Proxy
 
 import           GHC.Exts (IsList(..), IsString)
+import           GHC.TypeLits
 
 import           Crypto.Hash (hashWith, SHA256(..))
 import           Crypto.Number.Serialize (os2ip, i2ospOf_)
@@ -290,3 +293,118 @@ phraseToSeed mw dic passphrase =
   where
     sentence = toData $ mnemonicPhraseToString dic mw
     toData = String.toBytes String.UTF8
+
+
+-- -------------------------------------------------------------------------- --
+-- Mnemonic Sentence and Mnemonic Phrase
+-- -------------------------------------------------------------------------- --
+
+-- | Mnemonic Sentence is a list of 'WordIndex'.
+--
+-- This is the generic representation of a mnemonic phrase that can be used for
+-- transalating to a different dictionary (example: English to Japanese).
+--
+-- This is mainly used to convert from/to the 'Entropy' and for 'cardanoSlSeed'
+--
+newtype MnemonicSentence (mw :: Nat) = MnemonicSentence
+    { mnemonicSentenceToListN :: ListN mw WordIndex
+    }
+  deriving (Show, Eq, Ord, Typeable, NormalForm)
+instance ValidMnemonicSentence mw => IsList (MnemonicSentence mw) where
+    type Item (MnemonicSentence mw) = WordIndex
+    fromList = MnemonicSentence . fromMaybe (error "invalid mnemonic size") . ListN.toListN
+    toList = ListN.unListN . mnemonicSentenceToListN
+
+-- | Type Constraint to validate the given 'Nat' is valid for the supported
+-- 'MnemonicSentence'
+type ValidMnemonicSentence (mw :: Nat) =
+    ( KnownNat mw
+    , NatWithinBound Int mw
+    , Elem mw '[9, 12, 15, 18, 21, 24]
+    )
+
+-- | Human readable representation of a 'MnemonicSentence'
+--
+newtype MnemonicPhrase (mw :: Nat) = MnemonicPhrase
+    { mnemonicPhraseToListN :: ListN mw String
+    }
+  deriving (Show, Eq, Ord, Typeable, NormalForm)
+instance ValidMnemonicSentence mw => IsList (MnemonicPhrase mw) where
+    type Item (MnemonicPhrase mw) = String
+    fromList = fromMaybe (error "invalid mnemonic phrase") . mnemonicPhrase
+    toList = ListN.unListN . mnemonicPhraseToListN
+
+mnemonicPhrase :: forall mw . ValidMnemonicSentence mw => [String] -> Maybe (MnemonicPhrase mw)
+mnemonicPhrase l = MnemonicPhrase <$> ListN.toListN l
+{-# INLINABLE mnemonicPhrase #-}
+
+-- | check a given 'MnemonicPhrase' is valid for the given 'Dictionary'
+--
+checkMnemonicPhrase :: forall mw . ValidMnemonicSentence mw
+                    => Dictionary
+                    -> MnemonicPhrase mw
+                    -> Bool
+checkMnemonicPhrase dic (MnemonicPhrase ln) =
+    ListN.foldl' (\acc s -> (dictionaryTestWord dic s && acc)) True ln
+
+-- | convert the given 'MnemonicPhrase' to a generic 'MnemonicSentence'
+-- with the given 'Dictionary'.
+--
+-- This function assumes the 'Dictionary' and the 'MnemonicPhrase' are
+-- compatible (see 'checkMnemonicPhrase').
+--
+mnemonicPhraseToMnemonicSentence :: forall mw . ValidMnemonicSentence mw
+                                 => Dictionary
+                                 -> MnemonicPhrase mw
+                                 -> MnemonicSentence mw
+mnemonicPhraseToMnemonicSentence dic (MnemonicPhrase ln) = MnemonicSentence $
+    ListN.map (dictionaryWordToIndex dic) ln
+
+-- | convert the given generic 'MnemonicSentence' to a human readable
+-- 'MnemonicPhrase' targetting the language of the given 'Dictionary'.
+mnemonicSentenceToMnemonicPhrase :: forall mw . ValidMnemonicSentence mw
+                                 => Dictionary
+                                 -> MnemonicSentence mw
+                                 -> MnemonicPhrase mw
+mnemonicSentenceToMnemonicPhrase dic (MnemonicSentence ln) = MnemonicPhrase $
+    ListN.map (dictionaryIndexToWord dic) ln
+
+mnemonicPhraseToString :: forall mw . ValidMnemonicSentence mw
+                       => Dictionary
+                       -> MnemonicPhrase mw
+                       -> String
+mnemonicPhraseToString dic (MnemonicPhrase ln) = mconcat $
+    intersperse (dictionaryWordSeparator dic) (ListN.unListN ln)
+
+mnemonicSentenceToString :: forall mw . ValidMnemonicSentence mw
+                         => Dictionary
+                         -> MnemonicSentence mw
+                         -> String
+mnemonicSentenceToString dic = mnemonicPhraseToString dic
+                             . mnemonicSentenceToMnemonicPhrase dic
+
+-- | translate the given 'MnemonicPhrase' from one dictionary into another.
+--
+-- This function assumes the source dictionary is compatible with the given
+-- 'MnemonicPhrase' (see 'checkMnemonicPhrase')
+--
+translateTo :: forall mw . ValidMnemonicSentence mw
+            => Dictionary -- ^ source dictionary
+            -> Dictionary -- ^ destination dictionary
+            -> MnemonicPhrase mw
+            -> MnemonicPhrase mw
+translateTo dicSrc dicDst (MnemonicPhrase ln) = MnemonicPhrase $
+    ListN.map (dictionaryIndexToWord dicDst . dictionaryWordToIndex dicSrc) ln
+
+------------------------------------------------------------------------
+-- Helpers
+------------------------------------------------------------------------
+
+-- | convenient type level constraint to validate a given 'Nat' e is an elemnt
+-- of the list of 'Nat' l.
+type family Elem (e :: Nat) (l :: [Nat]) :: Constraint where
+    Elem e '[] = TypeError ('Text "offset: field "
+             ':<>: 'ShowType e
+             ':<>: 'Text " not elements of valids values")
+    Elem e (e ': _) = ()
+    Elem e (_ ': xs) = Elem e xs
